+++ comet/seq_channel.go
@@ -169,8 +169,9 @@ func (c *SeqChannel) AddConn(key string, conn *Connection) (*hlist.Element, erro
 // RemoveConn implements the Channel RemoveConn method.
 func (c *SeqChannel) RemoveConn(key string, e *hlist.Element) error {
 	c.mutex.Lock()
-	c.conn.Remove(e)
+	conn := c.conn.Remove(e)
 	c.mutex.Unlock()
+	close(conn.Buf)
 	ConnStat.IncrRemove()
 	glog.Infof("user_key:\"%s\" remove conn = %d", key, c.conn.Len())
 	return nil